bug fixes and xml update

+++ include/amp/engine.h
@@ -23,6 +23,7 @@
  */
 
 #include <sys/time.h>
+#include <sys/types.h>
 
 typedef struct amp_error_t amp_error_t;
 typedef struct amp_connection_t amp_connection_t;
@@ -38,11 +39,11 @@ amp_error_t *amp_engine_error(amp_engine_t *engine);
 
 // supplies engine with n bytes of input, engine returns number of
 // bytes consumed, or -1 if there was an error
-int amp_engine_input(amp_engine_t *engine, char *src, size_t n);
+ssize_t amp_engine_input(amp_engine_t *engine, char *src, size_t n);
 
 // supplies engine with n bytes of space for output, engine returns
 // number of bytes produced, or -1 if there was an error
-int amp_engine_output(amp_engine_t *engine, char *dst, size_t n);
+ssize_t amp_engine_output(amp_engine_t *engine, char *dst, size_t n);
 
 time_t amp_engine_tick(amp_engine_t *engine, time_t now);
